% Generated by build script
\name{gdc_projects}
\alias{gdc_projects}
\title{List GDC projects}
\usage{
gdc_projects()
}
\arguments{
\item{size}{Number of results (default 20, max 10000)}
\item{from}{Offset for pagination (default 0)}
}
\value{tibble: id, name, project_id, primary_site, disease_type,}
\description{List GDC projects}
